+++ autogpts/forge/.gitignore
@@ -172,3 +172,5 @@ agbenchmark_config/reports
 .vscode
 ig_*
 agent.db
+agbenchmark_config/updates.json
+agbenchmark_config/challenges_already_beaten.json
