+++ benchmark/agbenchmark/__main__.py
@@ -19,7 +19,11 @@ from .utils.data_types import AgentBenchmarkConfig
 
 BENCHMARK_START_TIME_DT = datetime.now(timezone.utc)
 BENCHMARK_START_TIME = BENCHMARK_START_TIME_DT.strftime("%Y-%m-%dT%H:%M:%S+00:00")
-TEMP_FOLDER_ABS_PATH = Path(os.path.dirname(os.path.abspath(__file__))) / "temp_folder"
+TEMP_FOLDER_ABS_PATH = Path.cwd() / "agbenchmark_config" / "temp_folder"
+CHALLENGES_ALREADY_BEATEN = (
+    Path.cwd() / "agbenchmark_config" / "challenges_already_beaten.json"
+)
+UPDATES_JSON_PATH = Path.cwd() / "agbenchmark_config" / "updates.json"
 
 
 def get_agent_benchmark_config() -> AgentBenchmarkConfig:
