+++ benchmark/agbenchmark/agent_api_interface.py
@@ -3,9 +3,8 @@ import os
 import time
 from typing import Any, Dict, Optional
 
-from agbenchmark.__main__ import TEMP_FOLDER_ABS_PATH
+from agbenchmark.__main__ import TEMP_FOLDER_ABS_PATH, UPDATES_JSON_PATH
 from agbenchmark.agent_interface import get_list_of_file_paths
-from agbenchmark.app import UPDATES_JSON_PATH
 from agbenchmark.utils.data_types import ChallengeData
 from agent_protocol_client import AgentApi, ApiClient, Configuration, TaskRequestBody
 from agent_protocol_client.models.step import Step
