+++ benchmark/agbenchmark/generate_test.py
@@ -10,6 +10,7 @@ from typing import Any, Dict, Optional
 
 import pytest
 
+from agbenchmark.__main__ import CHALLENGES_ALREADY_BEATEN, UPDATES_JSON_PATH
 from agbenchmark.agent_api_interface import append_updates_file
 from agbenchmark.utils.challenge import Challenge
 from agbenchmark.utils.data_types import AgentBenchmarkConfig, ChallengeData
@@ -50,7 +51,7 @@ def create_single_test(
         test_name = self.data.name
 
         try:
-            with open("challenges_already_beaten.json", "r") as f:
+            with open(CHALLENGES_ALREADY_BEATEN, "r") as f:
                 challenges_beaten_in_the_past = json.load(f)
         except:
             challenges_beaten_in_the_past = {}
@@ -213,4 +214,18 @@ def challenge_should_be_ignored(json_file):
     return "challenges/deprecated" in json_file or "challenges/library" in json_file
 
 
+def initialize_updates_file():
+    if os.path.exists(UPDATES_JSON_PATH):
+        # If the file already exists, overwrite it with an empty list
+        with open(UPDATES_JSON_PATH, "w") as file:
+            json.dump([], file, indent=2)
+        print("Initialized updates.json by overwriting with an empty array")
+    else:
+        # If the file doesn't exist, create it and write an empty list
+        with open(UPDATES_JSON_PATH, "w") as file:
+            json.dump([], file, indent=2)
+        print("Created updates.json and initialized it with an empty array")
+
+
+initialize_updates_file()
 generate_tests()
